Performance improvement
Now instead of trying to draw to an image, then converting it to a
texture, and then rendering the texture to the whole screen, I am
instead drawing stripes of the original textures to stripes in the
screen. It's essentially the same as the untextured variant of the
raycaster, only now I'm "getting the colors" from the textures,
instead of being flat.