actually performant raycaster wait wtf
in this iteration of the thing we ditched most of the rendering
code. drawing strips to the screen, while faithful to the original
technique, is slow in modern hardware because it needs to be
constantly communicating with the gpu, and lets not even talk about
pixel by pixel rendering.

to address this, the new rendering method renders each wall and
floor/ceiling tile as entire quads using opengl directly (through
raylib's apis). raycasting is still used to decide what walls can be
seen through a more efficient binary search method, but the rendering
is actually in 3d right now, which means we can move the view up and
down without having to do perspective correction manually.

on the one hand it makes things easy but on the other we are not
really doing raycasting anymore. sadly it is not a proper method for
2024, at least there's an earlier iteration of this project that used
proper raycasting, one can always go back to that to see how horribly
slow it is, and why this is better.